+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Added
 
+-   [band] support for nonlinear scales for strip, quantile, schedule, and
+    violin charts
 -   [examples] stacked bar chart with a time scale
 -   [core] support for specifying bandwidth for continuous scales, e.g. linear
     scales and time scales